pyproject.toml
[project]
name = "bet-mind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "grandalf>=0.8",
    "httpx>=0.28.1",
    "langchain-anthropic>=1.3.5",
    "langchain-community>=0.4.1",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.42.0",
]
